requirements.txt
celery==5.2.7
Django==4.1.6
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
Pillow==9.4.0
redis==4.5.1
psycopg2-binary==2.8.5
